awesome: Add autostart applications to set background wallpaer
To set the background the nitrogen application is used. This has to be
installed and should first select one background for each of the
screens. This is because in the rc.lua the nitrogen's restore command is
used and that restores the wallpapers to previously set ones.

Also, display compositor picom is started before setting background.