zsh: Add path for clangd (Manual installation)
Clangd is not available as an arch package (not even in AUR), so it is
downloaded manually and unpacked. A symbolic link clangd is created in
the /opt directory.

The process is the same for almost any application that is not available
in the arch repos. Download the archive, extract its contents, and and
set an symbolic link in /opt directory.