Update hack fonts to support more symbols
Some of the symbols used by nvim was not supported with the previous
version of the Hack nerd fonts. This was seen with the update of
bufferline plugin for neovim as they started to use nerd fonts recently
(march, 2023).

Just re-downloading the Hack fonts from replacing the fonts fixed the
problem. I'm guessing the symbols are newly added after they were
previously installed.