nvim: Install popular color-schemes
The most popular color-schemes were installed and tested out. I did like
all of them, but I was not completely satisfied with anything out of the
box. There something or other that was not so great.

The idea is to keep them installed, and configure them to make it
perfect. Until then the Gruvbox is left as the default color-scheme.

Additional changes:
  - Stating explicit theme for lualine is removed, it seems to adapt
  according to the chosen color-scheme.