nvim: Integrate mason and mason-lspconfig
Mason acts a package manager to install and manage LSPs. So, the
necessary LSPs can now be installed from within neovim on the fly.

Mason-lspconfig bridges between mason and nvim-lspconfig. The
nvim-lspconfig gives server configurations, and to configure the servers
installed through Mason, mason-lspconfig is recommended.

Mason-lspconfig is configured to automatically setup the servers
installed via Mason. Custom server configurations is expected to be
placed in `server-settings` directory with the file name same as the
server name. If custom configuration is not presetn the default settings
is used, where the default settings binds the lsp keybindings and
disables virtual text.