nvim: Fix breaking changes when updating to latest mason-lspconfig
The `.setup_handlers()` function was removed in v2.0.0. I used
`get_installed_servers()` to get the server names to set the extra
configs and enable the servers. The `automatic_enable` setting is set to
false, the servers are enabled after setting the extra configs.

The way to set extra setting for nvim-lspconfig has been changed with
from Nvim 0.11+. The `vim.lsp.config(<server>, cfg)` is supposed to be
over `lspconfig.<server>.setup(cfg)`.

Also, the repository has been moved to `mason-org`, and the repository
path is updated as suggested in the release notes.

Find the release notes for v2.0.0
[here](https://github.com/mason-org/mason-lspconfig.nvim/releases/tag/v2.0.0)