nvim: Change comment plugin's keybindings
Change the keybinding to reflect the keybindings convention.

The way the keys are binding now with same binding for `toggler` and
`opleader` make some features of the plugin unusable. But, I feel with
this binding all my use cases of the plugins are covered.

By having the same bindings the `opleader` bindings work on visual mode
with the same bindings. The downside is that in normal mode the `opleader`
features doesn't work anymore.